created functions to get access_token based on github env vars `github_pat` and `provider`
